Yet another progress commit

+++ rqt_multiplot/include/rqt_multiplot/MessageFieldItem.h
@@ -33,10 +33,10 @@ namespace rqt_multiplot {
   
     MessageFieldItem* getParent() const;
     size_t getNumChildren() const;
-    MessageFieldItem* getChild(int row) const;
+    MessageFieldItem* getChild(size_t row) const;
     MessageFieldItem* getChild(const QString& name) const;
     MessageFieldItem* getDescendant(const QString& path) const;
-    size_t getRow() const;
+    int getRow() const;
     size_t getNumColumns() const;
     const QString& getName() const;
     const variant_topic_tools::DataType& getDataType() const;
